+++ packages/react-magma-dom/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 2.3.5-alpha.8 (2021-02-17)
+
+**Note:** Version bump only for package react-magma-dom
+
+<<<<<<< HEAD
+
+### 2.3.5-alpha.5 (2021-02-12)
+
+**Note:** Version bump only for package react-magma-dom
+
+=======
+
+### 2.3.5-alpha.7 (2021-02-17)
+
+**Note:** Version bump only for package react-magma-dom
+
 ### 2.3.5-alpha.6 (2021-02-12)
 
 **Note:** Version bump only for package react-magma-dom
@@ -10,6 +26,8 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ### 2.3.5-alpha.5 (2021-02-12)
 
 **Note:** Version bump only for package react-magma-dom
+
+> > > > > > > dev
 
 ### 2.3.5-alpha.4 (2021-02-12)