+++ packages/react-magma-icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 1.0.9-alpha.8 (2021-02-17)
+
+**Note:** Version bump only for package react-magma-icons
+
+### 1.0.9-alpha.7 (2021-02-17)
+
+**Note:** Version bump only for package react-magma-icons
+
 ### 1.0.9-alpha.6 (2021-02-12)
 
 **Note:** Version bump only for package react-magma-icons